Breaks out useful testing utilities
